Create base API crate
