--- crates/bitwarden-api-api/src/apis/configuration.rs
@@ -1,48 +1,0 @@
-/*
- * Bitwarden Internal API
- *
- * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
- *
- * The version of the OpenAPI document: latest
- *
- * Generated by: https://openapi-generator.tech
- */
-
-#[derive(Debug, Clone)]
-pub struct Configuration {
-    pub base_path: String,
-    pub user_agent: Option<String>,
-    pub client: reqwest_middleware::ClientWithMiddleware,
-    pub basic_auth: Option<BasicAuth>,
-    pub oauth_access_token: Option<String>,
-    pub bearer_access_token: Option<String>,
-    pub api_key: Option<ApiKey>,
-}
-
-pub type BasicAuth = (String, Option<String>);
-
-#[derive(Debug, Clone)]
-pub struct ApiKey {
-    pub prefix: Option<String>,
-    pub key: String,
-}
-
-impl Configuration {
-    pub fn new() -> Configuration {
-        Configuration::default()
-    }
-}
-
-impl Default for Configuration {
-    fn default() -> Self {
-        Configuration {
-            base_path: "https://api.bitwarden.com".to_owned(),
-            user_agent: Some("OpenAPI-Generator/latest/rust".to_owned()),
-            client: reqwest_middleware::ClientBuilder::new(reqwest::Client::new()).build(),
-            basic_auth: None,
-            oauth_access_token: None,
-            bearer_access_token: None,
-            api_key: None,
-        }
-    }
-}
