+++ crates/bitwarden-test/src/api.rs
@@ -1,4 +1,4 @@
-use bitwarden_api_api::apis::configuration::Configuration;
+use bitwarden_api_api::Configuration;
 
 /// Helper for testing the Bitwarden API using wiremock.
 ///
@@ -14,10 +14,7 @@ pub async fn start_api_mock(mocks: Vec<wiremock::Mock>) -> (wiremock::MockServer
         base_path: server.uri(),
         user_agent: Some("test-agent".to_string()),
         client: reqwest::Client::new().into(),
-        basic_auth: None,
         oauth_access_token: None,
-        bearer_access_token: None,
-        api_key: None,
     };
 
     (server, config)
