+++ support/openapi-template/reqwest-trait/api_mod.mustache
@@ -1,153 +1,5 @@
-use std::error;
-use std::fmt;
-{{#withAWSV4Signature}}
-use aws_sigv4;
-{{/withAWSV4Signature}}
-
-#[derive(Debug, Clone)]
-pub struct ResponseContent<T> {
-    pub status: reqwest::StatusCode,
-    pub content: String,
-    pub entity: Option<T>,
-}
-
-#[derive(Debug)]
-pub enum Error<T> {
-    Reqwest(reqwest::Error),
-    {{#supportMiddleware}}
-    ReqwestMiddleware(reqwest_middleware::Error),
-    {{/supportMiddleware}}
-    Serde(serde_json::Error),
-    Io(std::io::Error),
-    ResponseError(ResponseContent<T>),
-    {{#withAWSV4Signature}}
-    AWSV4SignatureError(aws_sigv4::http_request::Error),
-    {{/withAWSV4Signature}}
-    {{#supportTokenSource}}
-    TokenSource(Box<dyn std::error::Error + Send + Sync>),
-    {{/supportTokenSource}}
-}
-
-impl <T> fmt::Display for Error<T> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let (module, e) = match self {
-            Error::Reqwest(e) => ("reqwest", e.to_string()),
-            {{#supportMiddleware}}
-            Error::ReqwestMiddleware(e) => ("reqwest-middleware", e.to_string()),
-            {{/supportMiddleware}}
-            Error::Serde(e) => ("serde", e.to_string()),
-            Error::Io(e) => ("IO", e.to_string()),
-            Error::ResponseError(e) => ("response", format!("status code {}", e.status)),
-            {{#withAWSV4Signature}}
-            Error::AWSV4SignatureError(e) => ("aws v4 signature", e.to_string()),
-            {{/withAWSV4Signature}}
-            {{#supportTokenSource}}
-            Error::TokenSource(e) => ("token source failure", e.to_string()),
-            {{/supportTokenSource}}
-        };
-        write!(f, "error in {}: {}", module, e)
-    }
-}
-
-impl <T: fmt::Debug> error::Error for Error<T> {
-    fn source(&self) -> Option<&(dyn error::Error + 'static)> {
-        Some(match self {
-            Error::Reqwest(e) => e,
-            {{#supportMiddleware}}
-            Error::ReqwestMiddleware(e) => e,
-            {{/supportMiddleware}}
-            Error::Serde(e) => e,
-            Error::Io(e) => e,
-            Error::ResponseError(_) => return None,
-            {{#withAWSV4Signature}}
-            Error::AWSV4SignatureError(_) => return None,
-            {{/withAWSV4Signature}}
-            {{#supportTokenSource}}
-            Error::TokenSource(e) => &**e,
-            {{/supportTokenSource}}
-        })
-    }
-}
-
-impl <T> From<reqwest::Error> for Error<T> {
-    fn from(e: reqwest::Error) -> Self {
-        Error::Reqwest(e)
-    }
-}
-
-{{#supportMiddleware}}
-impl<T> From<reqwest_middleware::Error> for Error<T> {
-    fn from(e: reqwest_middleware::Error) -> Self {
-        Error::ReqwestMiddleware(e)
-    }
-}
-
-{{/supportMiddleware}}
-impl <T> From<serde_json::Error> for Error<T> {
-    fn from(e: serde_json::Error) -> Self {
-        Error::Serde(e)
-    }
-}
-
-impl <T> From<std::io::Error> for Error<T> {
-    fn from(e: std::io::Error) -> Self {
-        Error::Io(e)
-    }
-}
-
-pub fn urlencode<T: AsRef<str>>(s: T) -> String {
-    ::url::form_urlencoded::byte_serialize(s.as_ref().as_bytes()).collect()
-}
-
-pub fn parse_deep_object(prefix: &str, value: &serde_json::Value) -> Vec<(String, String)> {
-    if let serde_json::Value::Object(object) = value {
-        let mut params = vec![];
-
-        for (key, value) in object {
-            match value {
-                serde_json::Value::Object(_) => params.append(&mut parse_deep_object(
-                    &format!("{}[{}]", prefix, key),
-                    value,
-                )),
-                serde_json::Value::Array(array) => {
-                    for (i, value) in array.iter().enumerate() {
-                        params.append(&mut parse_deep_object(
-                            &format!("{}[{}][{}]", prefix, key, i),
-                            value,
-                        ));
-                    }
-                },
-                serde_json::Value::String(s) => params.push((format!("{}[{}]", prefix, key), s.clone())),
-                _ => params.push((format!("{}[{}]", prefix, key), value.to_string())),
-            }
-        }
-
-        return params;
-    }
-
-    unimplemented!("Only objects are supported with style=deepObject")
-}
-
-/// Internal use only
-/// A content type supported by this client.
-#[allow(dead_code)]
-enum ContentType {
-    Json,
-    Text,
-    Unsupported(String)
-}
-
-impl From<&str> for ContentType {
-    fn from(content_type: &str) -> Self {
-        if content_type.starts_with("application") && content_type.contains("json") {
-            return Self::Json;
-        } else if content_type.starts_with("text/plain") {
-            return Self::Text;
-        } else {
-            return Self::Unsupported(content_type.to_string());
-        }
-    }
-}
+// Reexport base types from bitwarden-api-base for backwards compatibility
+pub use bitwarden_api_base::*;
 
 {{#apiInfo}}
 {{#apis}}
@@ -155,7 +7,10 @@ pub mod {{{classFilename}}};
 {{/apis}}
 {{/apiInfo}}
 
-pub mod configuration;
+// Reexport Configuration type from bitwarden-api-base for backwards compatibility
+pub mod configuration {
+    pub use bitwarden_api_base::Configuration;
+}
 
 {{#topLevelApiClient}}
 use std::sync::Arc;
@@ -189,7 +44,7 @@ pub struct ApiClientMock {
  {{/mockall}}
 
 impl ApiClient {
-    pub fn new(configuration: &Arc<configuration::Configuration>) -> Self {
+    pub fn new(configuration: &Arc<bitwarden_api_base::Configuration>) -> Self {
         Self::Real(ApiClientReal {
             {{#apiInfo}}
             {{#apis}}
